Fix some test failures in new HTTP server code

+++ rhizome.h
@@ -707,7 +707,7 @@ int rhizome_fetch_status_html(struct strbuf *b);
 int rhizome_fetch_has_queue_space(unsigned char log2_size);
 
 struct http_response_parts {
-  int code;
+  uint16_t code;
   char *reason;
   int64_t range_start;
   int64_t content_length;
